backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
Pillow>=10.0,<12.0
